+++ kensho_kenverters/convert_output.py
@@ -8,11 +8,11 @@ from typing import Any
 from kensho_kenverters.constants import (
     CATEGORY_KEY,
     DOCUMENT_CATEGORY_KEY,
+    ELEMENT_TITLE_CONTENT_CATEGORIES,
     LOCATIONS_KEY,
+    TABLE_CONTENT_CATEGORIES,
     TABLE_KEY,
-    TEXT_CONTENT_CATEGORIES,
     TEXT_KEY,
-    TITLE_CONTENT_CATEGORIES,
     AnnotationType,
     ContentCategory,
     TableType,
@@ -79,11 +79,18 @@ def _get_markdown_text(item: dict[str, Any]) -> str:
     elif item[CATEGORY_KEY] == ContentCategory.H2.value.lower():
         return "## " + item[TEXT_KEY]  # type: ignore[no-any-return]
     # Add ### to figure titles and table titles
-    elif item[CATEGORY_KEY] in (
-        ContentCategory.TABLE_TITLE.value.lower(),
-        ContentCategory.FIGURE_TITLE.value.lower(),
+    elif (
+        item[CATEGORY_KEY]
+        in [content_type.lower() for content_type in ELEMENT_TITLE_CONTENT_CATEGORIES]
+        or item[CATEGORY_KEY] == ContentCategory.H3.value.lower()
     ):
         return "### " + item[TEXT_KEY]  # type: ignore[no-any-return]
+    # Add #### to H4
+    elif item[CATEGORY_KEY] == ContentCategory.H4.value.lower():
+        return "#### " + item[TEXT_KEY]  # type: ignore[no-any-return]
+    # Add ##### to H5
+    elif item[CATEGORY_KEY] == ContentCategory.H5.value.lower():
+        return "##### " + item[TEXT_KEY]  # type: ignore[no-any-return]
     return item[TEXT_KEY]  # type: ignore[no-any-return]
 
 
@@ -97,14 +104,8 @@ def _create_segment(
     # DOCUMENT is just a head node
     if content.type == DOCUMENT_CATEGORY_KEY:
         return {}
-    # For texts and titles, add the text content and the category
-    elif content.type in TITLE_CONTENT_CATEGORIES | TEXT_CONTENT_CATEGORIES:
-        segment = {
-            CATEGORY_KEY: content.type.lower(),
-            TEXT_KEY: content.content,
-        }
     # For tables, use table cell structures read above
-    elif content.type == ContentCategory.TABLE.value:
+    elif content.type in TABLE_CONTENT_CATEGORIES:
         # Construct the table from cells
         table_cells = content.children
         # Drop tables with no cells
@@ -115,17 +116,23 @@ def _create_segment(
         if len(table) == 0:
             return {}
         segment = {
-            CATEGORY_KEY: ContentCategory.TABLE.value.lower(),
+            CATEGORY_KEY: content.type.lower(),
             TABLE_KEY: table,
             TEXT_KEY: table_to_markdown(table),
         }
     elif content.type == ContentCategory.TABLE_CELL.value:
         # Skip - already accounted for in tables
         return {}
+    # For texts and titles, add the text content and the category
+    elif content.type in [e.value for e in ContentCategory]:
+        segment = {
+            CATEGORY_KEY: content.type.lower(),
+            TEXT_KEY: content.content,
+        }
     else:
         raise TypeError(
             f"Content category must be in {[e.value for e in ContentCategory]}. "
-            "Found {content.type}"
+            f"Found {content.type}"
         )
     return segment